FuzzDist.Jepsen.JepSir (fuzz_dist v0.0.1-scrappy)

Generic Jepsen client for Elixir.

Jepsen creates operations on the Control node driven by property test :generators.

The operations are executed on the target node's client as an Elixir @behavior.

JepSir is strict in following a happy path. Unexpected or malformed messages crash or raise to invalidate the client in Jepsen.

Function names and @specs are the same as the Jepsen client protocol.

Types

antidote_return()

Specs

antidote_return() ::
  :ok
  | {:ok, term()}
  | {:error, :aborted}
  | {:error, :timeout}
  | {:error, String.t()}

jepsen_return()

Specs

jepsen_return() :: %{
  :type => :fail | :info | :ok,
  optional(:value) => term(),
  optional(:error) => :aborted | :timeout | String.t()
}

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.