FuzzDist.Telemetry (fuzz_dist v0.0.1-scrappy)

Telemetry integration.

Match :ok = :telemetry/all as telemetry is expected as part of test.

FuzzDist executes the following events:

[:fuzz_dist, :beam] - Executed on receipt of BEAM monitored message/signal.

Measurements

%{}

Metadata:

:event - The BEAM event

[:fuzz_dist, :client] - Executed on connection of Jepsen client ws connection

Measurements

%{}

Metadata:

:antidote_conn - AntidoteDB connection that client will use for all transactions

[:fuzz_dist, :g_set_read, :start] - Executed on receipt of Jepsen :read operation.

Measurements

:system_time - The system time

Metadata:

%{}

[:fuzz_dist, :g_set_read, :stop] - Executed on completion of read before return to Jepsen.

Measurements

:duration - Duration of the read.

Metadata:

reply returned by read, may contain keys:

:type - :ok, :fail, :info.

:value - Value returned by read.

:error - Supplemental error information.

[:fuzz_dist, :g_set_add, :start] - Executed on receipt of Jepsen :add operation.

Measurements

:system_time - The system time

Metadata:

value: - The value to be added

[:fuzz_dist, :g_set_add, :stop] - Executed on completion of add before return to Jepsen.

Measurements

:duration - Duration of the add.

Metadata:

reply returned by add, may contain keys:

:type - :ok, :fail, :info.

:error - Supplemental error information.

[:fuzz_dist, :pn_counter_read, :start] - Executed on receipt of Jepsen :read operation.

Measurements

:system_time - The system time

Metadata:

%{}

[:fuzz_dist, :pn_counter_read, :stop] - Executed on completion of read before return to Jepsen.

Measurements

:duration - Duration of the read.

Metadata:

reply returned by read, may contain keys:

:type - :ok, :fail, :info.

:value - Value returned by read.

:error - Supplemental error information.

[:fuzz_dist, :pn_counter_increment, :start] - Executed on receipt of Jepsen :increment operation.

Measurements

:system_time - The system time

Metadata:

value: - The [key value] to be incremented

[:fuzz_dist, :pn_counter_increment, :stop] - Executed on completion of :increment before return to Jepsen.

Measurements

:duration - Duration of the increment.

Metadata:

reply returned by increment, may contain keys:

:type - :ok, :fail, :info.

:error - Supplemental error information.

[:fuzz_dist, :pn_counter_decrement, :start] - Executed on receipt of Jepsen :decrement operation.

Measurements

:system_time - The system time

Metadata:

value: - The [key value] to be decremented

[:fuzz_dist, :pn_counter_decrement, :stop] - Executed on completion of :decrement before return to Jepsen.

Measurements

:duration - Duration of the decrement.

Metadata:

reply returned by decrement, may contain keys:

:type - :ok, :fail, :info.

:error - Supplemental error information.

[:fuzz_dist, :setup_primary, :start] - Executed on receipt of Jepsen :setup_primary operation.

Measurements

:system_time - The system time

Metadata:

nodes - List of (long) node names

[:fuzz_dist, :setup_primary, :stop] - Executed on completion of setup_primary/1 before return to Jepsen.

Measurements

:duration - Duration of the cluster configuration

Metadata:

%{}

TODO: add Prometheus collector.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

log_handler(event, measurements, meta, config)

start_link(opts)